refactor(idp-basic): Drop Relational prefix from store and GC types
Rename IdpRelationalEntityStore to IdpEntityStore, IdpRelationalGarbageCollector
to IdpGarbageCollector, and extract the IdpStore interface.
